+++ changelog.md
@@ -1,10 +1,15 @@
+## [1.5.1] - [2021-06-19]
+### Changed
+- Offenses are now called profiles
+- Violations are now called reports
+
 ## [1.5.0] - [2021-06-19]
 ### Added
 - WebSocket handler for FAGC events [1d639eb2f665e524b9a0822cdac3553f83a5ecb4]
 
 ## [1.4.0] - [2021-06-19]
 ### Added
-- Added back in the offense manager as stuff was hard without it [6f96ec45ca9e715a92b8b6e6f0813e449ce2c949]
+- Added back in the profile (offense) manager as stuff was hard without it [6f96ec45ca9e715a92b8b6e6f0813e449ce2c949]
 
 ## [1.2.0] - [2021-06-19]
 ### Added